+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @nextclaw/desktop
 
+## 0.0.88
+
+### Patch Changes
+
+- Updated dependencies
+  - @nextclaw/core@0.11.1
+  - @nextclaw/openclaw-compat@0.3.27
+  - @nextclaw/server@0.10.50
+  - nextclaw@0.15.8
+  - @nextclaw/channel-runtime@0.4.1
+  - @nextclaw/channel-plugin-dingtalk@0.2.15
+  - @nextclaw/channel-plugin-discord@0.2.15
+  - @nextclaw/channel-plugin-email@0.2.15
+  - @nextclaw/channel-plugin-mochat@0.2.15
+  - @nextclaw/channel-plugin-qq@0.2.15
+  - @nextclaw/channel-plugin-slack@0.2.15
+  - @nextclaw/channel-plugin-telegram@0.2.15
+  - @nextclaw/channel-plugin-wecom@0.2.15
+  - @nextclaw/channel-plugin-whatsapp@0.2.15
+
 ## 0.0.87
 
 ### Patch Changes